+++ packages/cms/lib/modules/arguments-widgets/index.js
@@ -1,23 +1,18 @@
 /**
  * A widget for displaying a list of arguments it's reactions, and a reaction form
- * Needs to be placed on a resource form 
+ * Needs to be placed on a resource form
  */
-const rp = require('request-promise');
-
 module.exports = {
   extend: 'openstad-widgets',
   label: 'Arguments old',
   alias: 'arguments',
   adminOnly: true,
   addFields: [
-  /*
-    Not being used for now,
-    but might in the future
     {
       name: 'ideaId',
       type: 'string',
       label: 'Idea ID (if empty it will try to fetch the ideaId from the URL)',
-    },*/
+    },
     {
       name: 'emptyPlaceholder',
       type: 'string',
@@ -86,13 +81,12 @@ module.exports = {
       {
         name: 'advanced',
         label: 'Advanced',
-        fields: ['replyingEnabled', 'votingEnabled']
+        fields: ['replyingEnabled', 'votingEnabled', 'ideaId']
       }
     ]);
 
 
      const superPushAssets = self.pushAssets;
-     //const auth = "Basic " + new Buffer("xxx:xxx#").toString("base64");
 
      self.pushAssets = function() {
        superPushAssets();
@@ -100,14 +94,35 @@ module.exports = {
        self.pushAsset('stylesheet', 'main', { when: 'always' });
      };
 
+    const superLoad = self.load;
+    self.load = async function (req, widgets, next) {
+      const promises = widgets.map(async (widget) => {
+        if (widget.ideaId) {
+          const resource = await self.apos.openstadApi.getResource(req, req.data.global.siteId, 'idea', widget.ideaId, {includeArguments: 1});
+          widget.ajaxError = null;
+          if (resource) {
+            widget.activeResource = resource
+            widget.activeResourceType = 'idea';
+            widget.activeResourceId = resource.id;
+          }
+        }
+      });
+
+      await Promise.all(promises);
+
+      return superLoad(req, widgets, next);
+    }
+
      var superOutput = self.output;
      self.output = function(widget, options) {
-       widget.ideaId =  options.activeResource ?  options.activeResource.id : false;
-       widget.activeResourceType = options.activeResourceType;
-       widget.activeResource = options.activeResource ?  options.activeResource : {};
-       widget.activeResourceId =  options.activeResource ?  options.activeResource.id : false;
+       if(!widget.ideaId || !widget.activeResource) {
+         widget.ideaId = options.activeResource ? options.activeResource.id : false;
+         widget.activeResourceType = options.activeResourceType;
+         widget.activeResource = options.activeResource ? options.activeResource : {};
+         widget.activeResourceId = options.activeResource ? options.activeResource.id : false;
+       }
+
        return superOutput(widget, options);
      };
-
-   },
+   }
 };